Cypress/E2E: Fix MM-T1752 enable account creation (#7429)
* fix and reorganize tests for authentication

* fix MM-T1752 enable account creation

* fix MM-T1770 for E20
