+++ e2e/cypress/integration/enterprise/auth_sso/mfa_authentication_spec.js
@@ -7,9 +7,9 @@
 // - Use element ID when selecting an element. Create one if none.
 // ***************************************************************
 
-// Group: @enterprise @system_console @authentication
+// Group: @enterprise @system_console @authentication @mfa
 
-import * as TIMEOUTS from '../../../../fixtures/timeouts';
+import * as TIMEOUTS from '../../../fixtures/timeouts';
 
 const authenticator = require('authenticator');
