+++ e2e/cypress/integration/enterprise/system_console/authentication_method_spec.js
@@ -7,7 +7,7 @@
 // - Use element ID when selecting an element. Create one if none.
 // ***************************************************************
 
-// Group: @enterpise @system_console @mfa
+// Group: @enterprise @system_console @mfa
 
 import ldapUsers from '../../../fixtures/ldap_users.json';
 import * as TIMEOUTS from '../../../fixtures/timeouts';
