+++ e2e/cypress/support/api/on_prem_default_config.json
@@ -17,7 +17,6 @@
         "ReadTimeout": 300,
         "WriteTimeout": 300,
         "IdleTimeout": 300,
-        "MaximumLoginAttempts": 10,
         "GoroutineHealthThreshold": -1,
         "GoogleDeveloperKey": "",
         "EnableOAuthServiceProvider": false,
